refactor: legal directory with terms, privacy and disclaimers tabs
