--- packages/curve-ui-kit/src/widgets/Disclaimer/index.ts
@@ -1,1 +1,0 @@
-export { Disclaimer } from './Disclaimer'
