+++ packages/curve-ui-kit/src/widgets/Legal/components/disclaimer-tabs/LlamaLend.tsx
@@ -6,8 +6,8 @@ import { t } from '@ui-kit/lib/i18n'
 import { type AppName, getInternalUrl, PAGE_DISCLAIMER } from '@ui-kit/shared/routes'
 import { RouterLink } from '@ui-kit/shared/ui/RouterLink'
 import { pushSearchParams } from '@ui-kit/utils/urls'
-import { List } from '../List'
-import { Header, Paragraph, Section, Title } from '../Section'
+import { List } from '../general/List'
+import { Header, Paragraph, Section, Title } from '../general/Section'
 
 export const LlamaLend = ({
   currentApp,
