+++ packages/curve-ui-kit/src/widgets/Legal/components/disclaimer-tabs/SCrvUsd.tsx
@@ -8,8 +8,8 @@ import { t } from '@ui-kit/lib/i18n'
 import { CRVUSD_ROUTES, getInternalUrl } from '@ui-kit/shared/routes'
 import { RouterLink } from '@ui-kit/shared/ui/RouterLink'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { List } from '../List'
-import { Header, Paragraph, Section } from '../Section'
+import { List } from '../general/List'
+import { Header, Paragraph, Section } from '../general/Section'
 
 const { Spacing } = SizesAndSpaces
