+++ packages/curve-ui-kit/src/widgets/Legal/components/general/Footer.tsx
