+++ packages/curve-ui-kit/src/widgets/Legal/components/general/Section.tsx
