+++ packages/curve-ui-kit/src/widgets/Legal/components/general/TabPanel.tsx
