+++ packages/curve-ui-kit/src/widgets/Legal/types/tabs.ts
@@ -1,0 +1,2 @@
+export type Tab = 'terms' | 'privacy' | 'disclaimers'
+export type DisclaimerTab = 'dex' | 'lend' | 'crvusd' | 'scrvusd'
